chore(harvest): relocate code under basic_knowledge/ and keep inputs in method_frame; update CI & outputs path [skip ci]
