+++ basic_knowledge/__init__.py
@@ -1,0 +1,7 @@
+"""Knowledge base tooling and artifacts."""
+
+from . import method_harvester
+
+__all__ = [
+    "method_harvester",
+]
