+++ basic_knowledge/_index/SUMMARY.md
@@ -1,6 +1,6 @@
 # Method Framework Knowledge Base
 
-_Last build: 2025-09-19T04:01:26+00:00_
+_Last build: 2025-09-19T05:55:48+00:00_
 
 **Total items:** 4
