+++ basic_knowledge/_index/extracted.jsonl
@@ -1,0 +1,4 @@
+{"item_id": "15c980c5c8986f1f1498dee5e5bea3cd1584e39b", "name": "Impact Effort Matrix Method", "aka": [], "type": "framework", "one_liner": "Plot work by impact and effort to surface quick wins and strategic bets.", "when_to_use": ["Quarterly planning", "Backlog grooming"], "inputs": [], "steps": ["List candidate initiatives", "Assess relative impact (high/low)", "Assess relative effort (high/low)", "Place each initiative in the quadrant grid"], "decision_rules": ["Execute high-impact, low-effort items first", "Re-evaluate high-effort, low-impact work"], "anti_patterns": [], "evidence_examples": [], "tags": ["prioritisation", "decision-making"], "sources": [{"file": "basic_knowledge/method_frame/_samples/prioritization.md", "anchor": "#impact-effort-matrix-method", "hash": "c66707fb81ade61c08b1f15f60eda987f4d7fea4", "file_sha256": "707b981d3e2183b8669862676e7d402c3a48bd240da3fc9c98af1767d4817248"}], "created_at": "2025-09-19T05:55:46+00:00", "updated_at": "2025-09-19T05:55:46+00:00"}
+{"item_id": "518a585e79bb67f3a459378ce47c70263ed98dc3", "name": "RICE Scoring Model", "aka": ["Reach", "Impact", "Confidence", "Effort"], "type": "heuristic", "one_liner": "Composite score for ranking initiatives by potential value and confidence.", "when_to_use": ["Product roadmap discussions"], "inputs": ["Reach estimate", "impact scale", "confidence percentage", "effort estimate"], "steps": ["Estimate reach over a fixed period", "Assign an impact score (3 = massive, 2 = high, 1 = medium, 0.5 = low)", "Assign confidence percentage to reflect certainty", "Estimate effort in person-months", "Calculate score = (Reach * Impact * Confidence) / Effort"], "decision_rules": ["Investigate items with low confidence before committing"], "anti_patterns": ["Inflating reach without data", "Ignoring effort estimates"], "evidence_examples": [], "tags": ["prioritisation", "product"], "sources": [{"file": "basic_knowledge/method_frame/_samples/prioritization.md", "anchor": "#rice-scoring-model", "hash": "b5f9fbfe52c3f384ff5bb8fc05bfd30b5915b42f", "file_sha256": "707b981d3e2183b8669862676e7d402c3a48bd240da3fc9c98af1767d4817248"}], "created_at": "2025-09-19T05:55:46+00:00", "updated_at": "2025-09-19T05:55:46+00:00"}
+{"item_id": "10f0021dec87225fc0753a842775f9c456f61783", "name": "5 Whys Method", "aka": ["Five Whys"], "type": "methodology", "one_liner": "Iteratively ask \"why\" to uncover the root cause of a problem.", "when_to_use": ["Incident reviews", "Customer escalations"], "inputs": [], "steps": ["Describe the problem clearly", "Ask \"why\" and capture the answer", "Repeat the questioning until the underlying cause is exposed"], "decision_rules": ["Stop once the team agrees the cause is actionable"], "anti_patterns": ["Stopping too early", "Blaming individuals without evidence"], "evidence_examples": ["Manufacturing line stoppage traced to a missed maintenance schedule"], "tags": ["rca", "problem-solving"], "sources": [{"file": "basic_knowledge/method_frame/_samples/root_cause_analysis.md", "anchor": "#5-whys-method", "hash": "bddf2b430ad23bca4855a3a87686c434c590dfc6", "file_sha256": "e859eb76deab61aba30b1d58128f042cad264d2fce48726cf267fbc0f27b0967"}], "created_at": "2025-09-19T05:55:46+00:00", "updated_at": "2025-09-19T05:55:46+00:00"}
+{"item_id": "0335c1800b5a9ed0595f829e6423d75e25220d2e", "name": "Fishbone Diagram Framework", "aka": ["Ishikawa Diagram", "Cause and Effect Diagram"], "type": "framework", "one_liner": "Visual map that categorises potential causes contributing to an effect.", "when_to_use": ["Cross-functional retrospectives", "Brainstorming workshops"], "inputs": ["Identified issue", "brainstorming team"], "steps": ["Define the effect at the head of the diagram", "Identify major cause categories (people, process, tools, environment)", "Brainstorm contributing factors for each category"], "decision_rules": ["Prioritise causes that appear across multiple categories"], "anti_patterns": ["Jumping to solutions before the diagram is populated"], "evidence_examples": [], "tags": ["rca", "facilitation"], "sources": [{"file": "basic_knowledge/method_frame/_samples/root_cause_analysis.md", "anchor": "#fishbone-diagram-framework", "hash": "cd5203c78eb8f11deb69d36723a8696dd7100a11", "file_sha256": "e859eb76deab61aba30b1d58128f042cad264d2fce48726cf267fbc0f27b0967"}], "created_at": "2025-09-19T05:55:46+00:00", "updated_at": "2025-09-19T05:55:46+00:00"}
