+++ basic_knowledge/_index/registry.json
@@ -237,6 +237,6 @@
   "metadata": {
     "created_at": "2025-09-19T04:01:21+00:00",
     "item_count": 4,
-    "updated_at": "2025-09-19T04:01:21+00:00"
+    "updated_at": "2025-09-19T05:55:47+00:00"
   }
 }
