+++ method_harvester/README.md
@@ -1,0 +1,4 @@
+# Package relocated
+
+The method harvester package moved to `basic_knowledge/method_harvester/`.
+Update imports to `basic_knowledge.method_harvester`.
