+++ scripts/README.md
@@ -1,0 +1,4 @@
+# Scripts relocated
+
+The harvester CLI now lives at `basic_knowledge/scripts/harvest_methods.py`.
+Use `python basic_knowledge/scripts/harvest_methods.py <command>` instead of the old path.
